admin dashboard

+++ routes/web.php
@@ -56,6 +56,7 @@ Route::middleware('auth')->group(function () {
 
 
 
+Route::get('/home',[DashboardController::class,'MultiAuth'])->name('AdminPanel.AdminDashboard');
 
 
 //search view
@@ -70,19 +71,21 @@ Route::get('dashboard/Layout/aboutus', function () {
 })->name('dashboard.Layout.aboutus');
 
 //Admin Route
-Route::get('AdminPanel/AdminDashboard',[DashboardController::class,'index'])->middleware('admin');
-Route::get('AdminPanel/ViewUser',[DashboardController::class,'user'])->name('AdminPanel.ViewUser');
-Route::get('AdminPanel/ViewMovie',[MovieController::class,'View'])->name('AdminPanel.ViewMovie');
-Route::get('AdminPanel/AddMovie',[DashboardController::class,'addMovie'])->name('AdminPanel.AddMovie');
-Route::get('AdminPanel/AddUser',[DashboardController::class,'addUser'])->name('AdminPanel.AddUser');
-Route::post('AdminPanel/store', [DashboardController::class, 'store'])->name('AdminPanel.store');
-Route::post('AdminPanel/storeMovie', [DashboardController::class, 'storeM'])->name('AdminPanel.storeM');
-Route::get('AdminPanel/EditMovie/{movie}/edit', [DashboardController::class, 'editMovie'])->name('AdminPanel.EditMovie');
-Route::put('AdminPanel/EditMovie/{movie}/update', [DashboardController::class, 'updateMovie'])->name('AdminPanel.updateM');
-Route::get('AdminPanel/EditUser/{user}/edit', [DashboardController::class, 'editUser'])->name('AdminPanel.EditUser');
-Route::put('AdminPanel/EditUser/{user}/update', [DashboardController::class, 'updateUser'])->name('AdminPanel.updateU');
-Route::delete('AdminPanel/ViewUser/{user}', [DashboardController::class, 'deleteU'])->name('AdminPanel.deleteU');
-Route::delete('AdminPanel/ViewMovie/{movie}', [DashboardController::class, 'deleteM'])->name('AdminPanel.deleteM');
+Route::get('AdminPanel/AdminDashboard',[DashboardController::class,'index'])->name('AdminPanel.AdminDashboard')->middleware('admin');
+Route::get('AdminPanel/ViewUser',[DashboardController::class,'user'])->name('AdminPanel.ViewUser')->middleware('admin');
+Route::get('AdminPanel/ViewMovie',[MovieController::class,'View'])->name('AdminPanel.ViewMovie')->middleware('admin');
+Route::get('AdminPanel/AddMovie',[DashboardController::class,'addMovie'])->name('AdminPanel.AddMovie')->middleware('admin');
+Route::get('AdminPanel/AddUser',[DashboardController::class,'addUser'])->name('AdminPanel.AddUser')->middleware('admin');
+Route::post('AdminPanel/store', [DashboardController::class, 'store'])->name('AdminPanel.store')->middleware('admin');
+Route::post('AdminPanel/storeMovie', [DashboardController::class, 'storeM'])->name('AdminPanel.storeM')->middleware('admin');
+Route::get('AdminPanel/EditMovie/{movie}/edit', [DashboardController::class, 'editMovie'])->name('AdminPanel.EditMovie')->middleware('admin');
+Route::put('AdminPanel/EditMovie/{movie}/update', [DashboardController::class, 'updateMovie'])->name('AdminPanel.updateM')->middleware('admin');
+Route::get('AdminPanel/EditUser/{user}/edit', [DashboardController::class, 'editUser'])->name('AdminPanel.EditUser')->middleware('admin');
+Route::put('AdminPanel/EditUser/{user}/update', [DashboardController::class, 'updateUser'])->name('AdminPanel.updateU')->middleware('admin');
+Route::delete('AdminPanel/ViewUser/{user}', [DashboardController::class, 'deleteU'])->name('AdminPanel.deleteU')->middleware('admin');
+Route::delete('AdminPanel/ViewMovie/{movie}', [DashboardController::class, 'deleteM'])->name('AdminPanel.deleteM')->middleware('admin');
+Route::get('AdminPanel/Profile',[DashboardController::class,'profile'])->name('AdminPanel.Profile')->middleware('admin');
+Route::put('AdminPanel/profileUpdate', [DashboardController::class, 'updateProfile'])->name('AdminPanel.profileUpdate');
 
 require __DIR__.'/auth.php';
